+++ clients/admin-ui/src/features/user-management/UserManagementTable.tsx
@@ -60,6 +60,7 @@ const UserManagementTable: React.FC<UsersTableProps> = () => {
         <Thead>
           <Tr>
             <Th pl={0}>Username</Th>
+            <Th pl={0}>Email</Th>
             <Th pl={0}>First Name</Th>
             <Th pl={0}>Last Name</Th>
             <Th pl={0}>Permissions</Th>