+++ clients/admin-ui/src/types/api/models/UserResponse.ts
@@ -9,6 +9,9 @@ export type UserResponse = {
   id: string;
   username: string;
   created_at: string;
+  email_address?: string;
   first_name?: string;
   last_name?: string;
+  disabled?: boolean;
+  disabled_reason?: string;
 };